+++ scanpipe/management/commands/graph.py
@@ -21,15 +21,18 @@
 # Visit https://github.com/nexB/scancode.io for support and download.
 
 import subprocess
+import sys
 
 from django.core.management import CommandError
 from django.core.management.base import BaseCommand
 
+from scanpipe.management.commands import scanpipe_app_config
 from scanpipe.pipelines import PipelineGraph
 from scanpipe.pipelines import get_pipeline_class
+from scanpipe.pipelines import get_pipeline_doc
 
 
-def graphviz_installed():
+def is_graphviz_installed():
     exitcode, _ = subprocess.getstatusoutput("which dot")
     if exitcode == 0:
         return True
@@ -43,18 +46,39 @@ class Command(BaseCommand):
         parser.add_argument(
             "args",
             metavar="pipelines",
-            nargs="+",
+            nargs="*",
             help="One or more pipeline locations.",
+        )
+        parser.add_argument(
+            "--list",
+            action="store_true",
+            help="Display a list of all available pipelines.",
         )
         parser.add_argument("--output", help="Output directory location.")
 
     def handle(self, *pipelines, **options):
-        if not graphviz_installed():
+        if options["list"]:
+            for location, _ in scanpipe_app_config.pipelines:
+                self.stdout.write("- " + self.style.SUCCESS(location))
+                self.stdout.write(get_pipeline_doc(location), ending="\n\n")
+            sys.exit(0)
+
+        if not is_graphviz_installed():
             raise CommandError("Graphviz is not installed.")
+
+        if not pipelines:
+            self.stderr.write(self.style.ERROR("The pipelines argument is required."))
+            sys.exit(1)
 
         outputs = []
         for pipeline_location in pipelines:
-            pipeline_class = get_pipeline_class(pipeline_location)
+            try:
+                pipeline_class = get_pipeline_class(pipeline_location)
+            except ModuleNotFoundError:
+                self.stderr.write(
+                    self.style.ERROR(f"{pipeline_location} is not valid.")
+                )
+                sys.exit(1)
             pipeline_graph = PipelineGraph(pipeline_class)
             outputs.append(self.generate_graph(pipeline_graph, options.get("output")))